+++ Changes.md
@@ -5,6 +5,7 @@
 - Remove deprecated APIs
 - Use the serde serialization APIs directly, instead of using the bytecode
   serializer. Serialization will be about `2x` faster
+- Fix bug in `SchemaLike::from_value` with incorrect strategy deserialization
 
 ## 0.9.1